refactor(web): authenticate を FSD 構成へ再編し機能単位バレルに細分化
- 共有ドメイン（フォーム・バリデーション・エラーメッセージ・セッション・Turnstile）を
  entities/auth スライス（ui/model/lib）へ集約し、公開面を index.ts に閉じる
- features/authenticate を login / signup / logout の機能単位ディレクトリ＋専用バレルに再編
- 重い index.ts と軽量 ui.ts の二分を廃止。レイアウトは logout バレルを参照する形にし、
  SWR 系フック（use-sync-external-store）の巻き込みを機能バレル単位で自然に回避する

Refs #866
