+++ application/apps/web/src/client/components/ui/layout/app-header/app-header.stories.tsx
@@ -3,7 +3,7 @@ import { expect } from 'storybook/test'
 import { vi } from 'vitest'
 import AppHeader from './index'
 
-vi.mock('@/client/features/authenticate/hooks/use-logout', () => ({
+vi.mock('@/client/features/authenticate/logout/use-logout', () => ({
   default: vi.fn(() => ({
     handleLogout: vi.fn(),
     isLoading: false,
