+++ application/apps/web/src/client/entities/auth/index.ts
@@ -1,0 +1,9 @@
+export { AuthenticateForm, type AuthenticateFormBaseProps } from './ui/authenticate-form'
+export {
+  AUTH_ERROR_MESSAGES,
+  resolveLoginErrorMessage,
+  resolveSignupErrorMessage,
+} from './lib/error-message'
+export { resolveTurnstileSiteKey } from './lib/turnstile'
+export { default as useSession, SESSION_SWR_KEY } from './model/use-session'
+export { type AuthenticateErrors, authenticateFormSchema } from './model/validation'
