--- application/apps/web/src/client/features/authenticate/index.ts
@@ -1,16 +1,0 @@
-export {
-  AuthenticateForm,
-  type AuthenticateFormBaseProps,
-  LogoutButton,
-  SidebarLogoutButton,
-} from './ui'
-export {
-  AUTH_ERROR_MESSAGES,
-  resolveLoginErrorMessage,
-  resolveSignupErrorMessage,
-} from './model/error-message'
-export { resolveTurnstileSiteKey } from './model/turnstile'
-export { default as useLogin } from './hooks/use-login'
-export { default as useSession, SESSION_SWR_KEY } from './hooks/use-session'
-export { default as useSignup } from './hooks/use-signup'
-export { type AuthenticateErrors } from './model/validation'
