+++ application/apps/web/src/client/features/authenticate/login/index.ts
@@ -1,0 +1,1 @@
+export { default as useLogin } from './use-login'
