+++ application/apps/web/src/client/features/authenticate/logout/index.ts
@@ -1,0 +1,2 @@
+export { default as LogoutButton } from './logout-button'
+export { default as SidebarLogoutButton } from './sidebar-logout-button'
