+++ application/apps/web/src/client/features/authenticate/logout/logout-button.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { expect, fn, userEvent } from 'storybook/test'
 import { vi } from 'vitest'
-import useLogout from '@/client/features/authenticate/hooks/use-logout'
 import LogoutButton from './logout-button'
+import useLogout from './use-logout'
 
-vi.mock('@/client/features/authenticate/hooks/use-logout', () => ({
+vi.mock('./use-logout', () => ({
   default: vi.fn(),
 }))
