+++ application/apps/web/src/client/features/authenticate/logout/sidebar-logout-button.tsx
@@ -1,5 +1,5 @@
 import { SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/client/components/shadcn/sidebar'
-import useLogout from '@/client/features/authenticate/hooks/use-logout'
+import useLogout from './use-logout'
 
 export default function SidebarLogoutButton() {
   const { handleLogout, isLoading } = useLogout()
