+++ application/apps/web/src/client/features/authenticate/logout/use-logout.ts
