+++ application/apps/web/src/client/features/authenticate/signup/index.ts
@@ -1,0 +1,1 @@
+export { default as useSignup } from './use-signup'
