--- application/apps/web/src/client/features/authenticate/ui.ts
@@ -1,6 +1,0 @@
-// index.ts はSWR系フック（useSession 等）も束ねており、レイアウトのStorybookから読むと
-// use-sync-external-store の取り込みでブラウザバンドルが壊れる。UIだけを使う側はこの軽量な
-// 公開口から取り込むことでフック群の巻き込みを避ける。
-export { AuthenticateForm, type AuthenticateFormBaseProps } from './components/authenticate-form'
-export { default as LogoutButton } from './components/logout-button'
-export { default as SidebarLogoutButton } from './components/sidebar-logout-button'
